implemented additional test-cases for preview-form view
