+++ src/Sulu/Bundle/AdminBundle/Resources/js/views/Form/tests/Form.test.js
@@ -289,104 +289,6 @@ test('Should add items defined in ToolbarActions to Toolbar', () => {
     ]);
 });
 
-test('Should initialize preview sidebar', () => {
-    const withSidebar = require('../../../containers/Sidebar/withSidebar');
-    const Form = require('../Form').default;
-    const sidebarFunction = findWithHighOrderFunction(withSidebar, Form);
-    const ResourceStore = require('../../../stores/ResourceStore').default;
-    const resourceStore = new ResourceStore('snippet', 1);
-    resourceStore.loading = false;
-
-    const route = {
-        options: {
-            formKey: 'snippets',
-            preview: 'nodeType == 1',
-            toolbarActions: [],
-        },
-    };
-    const router = {
-        addUpdateRouteHook: jest.fn(),
-        bind: jest.fn(),
-        route,
-        attributes: {},
-    };
-
-    const form = mount(<Form resourceStore={resourceStore} route={route} router={router} />);
-    // trigger evaluation
-    sidebarFunction.call(form.instance());
-
-    setTimeout(() => {
-        const sidebarConfig = sidebarFunction.call(form.instance());
-        expect(sidebarConfig.view).toEqual('sulu_preview.preview');
-        expect(sidebarConfig.sizes).toEqual(['medium', 'large']);
-        expect(sidebarConfig.props.router).toEqual(router);
-        expect(sidebarConfig.props.formStore).toBeDefined();
-    }, 0);
-});
-
-test('Should not initialize preview sidebar when expression evaluates to false', () => {
-    const withSidebar = require('../../../containers/Sidebar/withSidebar');
-    const Form = require('../Form').default;
-    const sidebarFunction = findWithHighOrderFunction(withSidebar, Form);
-    const ResourceStore = require('../../../stores/ResourceStore').default;
-    const resourceStore = new ResourceStore('snippet', 1);
-    resourceStore.loading = false;
-
-    const route = {
-        options: {
-            formKey: 'snippets',
-            toolbarActions: [],
-            preview: 'nodeType == 2',
-        },
-    };
-    const router = {
-        addUpdateRouteHook: jest.fn(),
-        bind: jest.fn(),
-        route,
-        attributes: {},
-    };
-
-    const form = mount(<Form resourceStore={resourceStore} route={route} router={router} />);
-    // trigger evaluation
-    sidebarFunction.call(form.instance());
-
-    setTimeout(() => {
-        const sidebarConfig = sidebarFunction.call(form.instance());
-        expect(sidebarConfig).toEqual(null);
-    }, 0);
-});
-
-test('Should not initialize preview sidebar when option is not set', () => {
-    const withSidebar = require('../../../containers/Sidebar/withSidebar');
-    const Form = require('../Form').default;
-    const sidebarFunction = findWithHighOrderFunction(withSidebar, Form);
-    const ResourceStore = require('../../../stores/ResourceStore').default;
-    const resourceStore = new ResourceStore('snippet', 1);
-    resourceStore.loading = false;
-
-    const route = {
-        options: {
-            formKey: 'snippets',
-            toolbarActions: [],
-        },
-    };
-    const router = {
-        addUpdateRouteHook: jest.fn(),
-        bind: jest.fn(),
-        route,
-        attributes: {},
-    };
-
-    const form = mount(<Form resourceStore={resourceStore} route={route} router={router} />);
-    // trigger evaluation
-    sidebarFunction.call(form.instance());
-
-    setTimeout(() => {
-        const sidebarConfig = sidebarFunction.call(form.instance());
-        expect(sidebarConfig).toEqual(null);
-    }, 0);
-});
-
 test('Should not add PublishIndicator if no publish status is available', () => {
     const withToolbar = require('../../../containers/Toolbar/withToolbar');
     const Form = require('../Form').default;
