+++ src/Sulu/Bundle/AdminBundle/Resources/js/views/PreviewForm/tests/PreviewForm.test.js
@@ -56,6 +56,37 @@ test('Should render Form view', () => {
     )).toMatchSnapshot();
 });
 
+test('Should initialize preview sidebar per default when previewCondition is not set', () => {
+    const resourceStore = new ResourceStore('snippet', 1);
+
+    const route: Route = ({
+        options: {},
+    }: any);
+    const router = ({
+        route: route,
+    }: any);
+
+    // require preview form to trigger call of withSidebar mock and retrieve passed function
+    const PreviewForm = (require('../PreviewForm').default: any);
+    const withSidebar = (require('../../../containers/Sidebar/withSidebar'): any);
+    const Form = (require('../../Form'): any);
+    const sidebarFunction = findWithHighOrderFunction(withSidebar, Form);
+
+    // mount PreviewForm and call function that was passed to withSidebar
+    const previewForm = mount(<PreviewForm locales={[]} resourceStore={resourceStore} route={route} router={router} />);
+    const sidebarConfig = sidebarFunction.call(previewForm.instance());
+
+    // check if function that was passed to withSidebar returns the correct SidebarConfig
+    expect(sidebarConfig.view).toEqual('sulu_preview.preview');
+    expect(sidebarConfig.sizes).toEqual(['medium', 'large']);
+    expect(sidebarConfig.props.router).toEqual(router);
+    expect(sidebarConfig.props.formStore).toBeDefined();
+
+    // check if evalSync was called with correct parameters during function call
+    const jexl = (require('jexl'): any);
+    expect(jexl.evalSync).not.toBeCalled();
+});
+
 test('Should initialize preview sidebar when previewCondition evaluates to true', () => {
     const resourceStore = new ResourceStore('snippet', 1);
 
@@ -87,4 +118,34 @@ test('Should initialize preview sidebar when previewCondition evaluates to true'
     // check if evalSync was called with correct parameters during function call
     const jexl = (require('jexl'): any);
     expect(jexl.evalSync).toBeCalledWith( 'nodeType == 1', {testKey: 'test-value'});
+});
+
+test('Should not initialize preview sidebar when previewCondition evaluates to true', () => {
+    const resourceStore = new ResourceStore('snippet', 1);
+
+    const route: Route = ({
+        options: {
+            previewCondition: 'nodeType == 2',
+        },
+    }: any);
+    const router = ({
+        route: route,
+    }: any);
+
+    // require preview form to trigger call of withSidebar mock and retrieve passed function
+    const PreviewForm = (require('../PreviewForm').default: any);
+    const withSidebar = (require('../../../containers/Sidebar/withSidebar'): any);
+    const Form = (require('../../Form'): any);
+    const sidebarFunction = findWithHighOrderFunction(withSidebar, Form);
+
+    // mount PreviewForm and call function that was passed to withSidebar
+    const previewForm = mount(<PreviewForm locales={[]} resourceStore={resourceStore} route={route} router={router} />);
+    const sidebarConfig = sidebarFunction.call(previewForm.instance());
+
+    // check if function that was passed to withSidebar returns the correct SidebarConfig
+    expect(sidebarConfig).toEqual(null);
+
+    // check if evalSync was called with correct parameters during function call
+    const jexl = (require('jexl'): any);
+    expect(jexl.evalSync).toBeCalledWith( 'nodeType == 2', {testKey: 'test-value'});
 });
